HW5_P2 #9 final product

+++ .ccls-cache/@home@runner@HomeworkJakeThompson/HW5_P2@main5_2.cpp
@@ -36,7 +36,7 @@ int main()
   }
   cout << "The sum of valid data is: " << sumData << endl;
   cout << "The number of valid data points is: " << counter << endl;
-  cout << "The average of the valid data is: " << (sumData/counter) << endl;
+  cout << "The average of the valid data is: " << (static_cast<double>(sumData)/counter) << endl;
 
   return 0;
 }
